% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_languages.R
\docType{data}
\name{ec_languages}
\alias{ec_languages}
\title{Catalogue of East Caucasian languages}
\format{
A data frame with 56 rows and 10 variables:
\describe{
  \item{language}{-}
  \item{x}{-}
  \item{y}{-}
  \item{aff}{-}
  \item{family}{-}
  \item{gltc_lang}{-}
  \item{lang_col}{-}
  \item{aff_col}{-}
  \item{abbr}{-}
  \item{feature}{-}
}
}
\usage{
ec_languages
}
\description{
A dataset containes metadata on East Caucasian languages that helps
RCaucTile to function.
}
\keyword{datasets}
