% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_tile_map.R
\name{ec_tile_map}
\alias{ec_tile_map}
\title{Draw a grid tile map for East Caucasian languages.}
\usage{
ec_tile_map(
  data = NULL,
  feature_column = "feature",
  title = NULL,
  title_position = "left",
  fill_by = "feature",
  abbreviation = TRUE
)
}
\arguments{
\item{data}{Could be \code{NULL}, then it will print the language template. Otherwise should be a dataframe with language column annotated with some feature. Default value is \code{NULL}.}

\item{feature_column}{Character vector of length 1 which specifies the column in dataframe that contains annotation for the feature to color the language template with.}

\item{title}{Character vector of length 1, which specifies the title of the plot.}

\item{title_position}{Character vector of length 1, which specifies the title's position. Possible values are \code{left}, \code{center}, and \code{right}. Default value is \code{left}.}

\item{fill_by}{Character vector of length 1 which specifies, whether internal part of the rectangular should be colored by the language (value \code{language}) or by the feature (value \code{feature}). Default value is \code{feature}.}

\item{abbreviation}{Logical variable that specifies, whether use abbreviations for languages specified in the package. Default value is \code{TRUE}.}
}
\value{
a `ggplot2` object
}
\description{
Draw a grid tile map for East Caucasian languages.
}
\examples{
ec_tile_map()
ec_tile_map(ec_languages,
            feature_column = "morning_greetings",
            title = "Morning greetings (Naccarato, Verhees 2021)")

}
