% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_languages.R
\docType{data}
\name{ec_languages}
\alias{ec_languages}
\title{Catalogue of East Caucasian languages}
\format{
A data frame with 56 rows and 10 variables:
\describe{
  \item{language}{- hi}
  \item{x}{- hi}
  \item{y}{- hi}
  \item{affiliation}{- hi}
  \item{family}{- hi}
  \item{glottocode}{- hi}
  \item{language_color}{- hi}
  \item{affiliation_color}{- hi}
  \item{abbreviation}{- hi}
}
}
\usage{
ec_languages
}
\description{
A dataset containes metadata on East Caucasian languages that helps
RCaucTile to function.
}
\keyword{datasets}
