% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_languages.R
\docType{data}
\name{ec_languages}
\alias{ec_languages}
\title{Catalogue of East Caucasian languages}
\format{
A data frame with 56 rows and 11 variables:
\describe{
  \item{language}{- language name}
  \item{x}{- suggested x coordinate for the tile map}
  \item{y}{- suggested y coordinate for the tile map}
  \item{branch}{- language branch}
  \item{family}{- language family}
  \item{glottocode}{- language glottocode}
  \item{language_color}{- suggested language color}
  \item{branch_color}{- suggested branch color}
  \item{abbreviation}{- suggested abbreviation}
  \item{morning_greetings}{This variable contains values from the \href{https://lingconlab.ru/tald/005_morning_greetings.html}{"Morning Greetings" chapter} (Naccarato, Verhees 2021) from the Typological Atlas of the Languages of Daghestan. The languages of Daghestan can be classified into three groups according to whether they feature morning greetings including questions about the night’s rest (value `Did you wake up?`), based on the combination of concepts like “morning” and “good” (value `Good morning`), and both stratagies (value `Both`).}
  \item{consonant_inventory_size}{This variable contains consonant inventory sizes based on \href{https://lingconlab.ru/tald/phonology.html}{"Phonology" chapter} (Moroz 2021) from the Typological Atlas of the Languages of Daghestan.}
}
}
\usage{
ec_languages
}
\description{
A dataset containes metadata on East Caucasian languages that helps
RCaucTile to function.
}
\keyword{datasets}
