% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_tile_map.R
\name{ec_tile_map}
\alias{ec_tile_map}
\title{Draw a grid tile map for East Caucasian languages.}
\usage{
ec_tile_map(
  data = NULL,
  feature_column = "feature",
  title = NULL,
  fill_by = "language",
  abbreviation = TRUE
)
}
\arguments{
\item{data}{Could be NULL, then it will print the language template. Otherwise should be a dataframe with language column annotated with some feature. Default value is NULL.}

\item{feature_column}{Character vector of length 1 which specifies the column in dataframe that contains annotation for the feature to color the language template with.}

\item{title}{Character vector of length 1 which specifies the title of the plot.}

\item{fill_by}{Character vector of length 1 which specifies, whether internal part of the rectangular should be colored by the language (value "language") or by the feature (value "feature"). Default value is "language".}

\item{abbreviation}{Logical variable that specifies, whether use abbreviations for languages specified in the package.}
}
\value{
a `ggplot2` object
}
\description{
Draw a grid tile map for East Caucasian languages.
}
\examples{
ec_tile_map()
ec_tile_map(ec_languages,
            feature_column = "morning_greetings",
            fill_by = "feature",
            title = "Morning greetings (Naccarato, Verhees 2021)")

}
